Splitting particleset tests into two files
One for the creation of particlesets, and one focussed on the pset.execute behaviour
