+++ tests/v4/test_particleset_execute.py
@@ -1,0 +1,136 @@
+import numpy as np
+import pytest
+
+from parcels import (
+    AdvectionEE,
+    Field,
+    FieldSet,
+    Particle,
+    ParticleSet,
+    StatusCode,
+    UXPiecewiseConstantFace,
+    VectorField,
+    xgcm,
+)
+from parcels._datasets.structured.generic import datasets as datasets_structured
+from parcels._datasets.unstructured.generic import datasets as datasets_unstructured
+from parcels.uxgrid import UxGrid
+from parcels.xgrid import XGrid
+
+
+@pytest.fixture
+def fieldset() -> FieldSet:
+    ds = datasets_structured["ds_2d_left"]
+    grid = XGrid(xgcm.Grid(ds))
+    U = Field("U", ds["U (A grid)"], grid, mesh_type="flat")
+    V = Field("V", ds["V (A grid)"], grid, mesh_type="flat")
+    return FieldSet([U, V])
+
+
+def test_pset_remove_particle_in_kernel(fieldset, npart=100):
+    pset = ParticleSet(fieldset, lon=np.linspace(0, 1, npart), lat=np.linspace(1, 0, npart))
+
+    def DeleteKernel(particle, fieldset, time):  # pragma: no cover
+        if particle.lon >= 0.4:
+            particle.delete()
+
+    pset.execute(pset.Kernel(DeleteKernel), runtime=np.timedelta64(1, "s"), dt=np.timedelta64(1, "s"))
+    assert pset.size == 40
+
+
+def test_pset_stop_simulation(fieldset):
+    pset = ParticleSet(fieldset, lon=0, lat=0, pclass=Particle)
+
+    def Delete(particle, fieldset, time):  # pragma: no cover
+        if time >= fieldset.U.time[0].values + np.timedelta64(4, "s"):
+            return StatusCode.StopExecution
+
+    pset.execute(Delete, dt=np.timedelta64(1, "s"), runtime=np.timedelta64(21, "s"))
+    assert pset[0].time == fieldset.U.time[0].values + np.timedelta64(4, "s")
+
+
+@pytest.mark.parametrize("verbose_progress", [True, False])
+def test_uxstommelgyre_pset_execute(verbose_progress):
+    ds = datasets_unstructured["stommel_gyre_delaunay"]
+    grid = UxGrid(grid=ds.uxgrid, z=ds.coords["nz"])
+    U = Field(
+        name="U",
+        data=ds.U,
+        grid=grid,
+        mesh_type="spherical",
+        interp_method=UXPiecewiseConstantFace,
+    )
+    V = Field(
+        name="V",
+        data=ds.V,
+        grid=grid,
+        mesh_type="spherical",
+        interp_method=UXPiecewiseConstantFace,
+    )
+    P = Field(
+        name="P",
+        data=ds.p,
+        grid=grid,
+        mesh_type="spherical",
+        interp_method=UXPiecewiseConstantFace,
+    )
+    UV = VectorField(name="UV", U=U, V=V)
+    fieldset = FieldSet([UV, UV.U, UV.V, P])
+    pset = ParticleSet(
+        fieldset,
+        lon=[30.0],
+        lat=[5.0],
+        depth=[50.0],
+        time=[np.timedelta64(0, "s")],
+        pclass=Particle,
+    )
+    pset.execute(
+        runtime=np.timedelta64(10, "m"),
+        dt=np.timedelta64(60, "s"),
+        pyfunc=AdvectionEE,
+        verbose_progress=verbose_progress,
+    )
+
+
+@pytest.mark.xfail(reason="Output file not implemented yet")
+def test_uxstommelgyre_pset_execute_output():
+    ds = datasets_unstructured["stommel_gyre_delaunay"]
+    grid = UxGrid(grid=ds.uxgrid, z=ds.coords["nz"])
+    U = Field(
+        name="U",
+        data=ds.U,
+        grid=grid,
+        mesh_type="spherical",
+        interp_method=UXPiecewiseConstantFace,
+    )
+    V = Field(
+        name="V",
+        data=ds.V,
+        grid=grid,
+        mesh_type="spherical",
+        interp_method=UXPiecewiseConstantFace,
+    )
+    P = Field(
+        name="P",
+        data=ds.p,
+        grid=grid,
+        mesh_type="spherical",
+        interp_method=UXPiecewiseConstantFace,
+    )
+    UV = VectorField(name="UV", U=U, V=V)
+    fieldset = FieldSet([UV, UV.U, UV.V, P])
+    pset = ParticleSet(
+        fieldset,
+        lon=[30.0],
+        lat=[5.0],
+        depth=[50.0],
+        time=[0.0],
+        pclass=Particle,
+    )
+    output_file = pset.ParticleFile(
+        name="stommel_uxarray_particles.zarr",  # the file name
+        outputdt=np.timedelta64(5, "m"),  # the time step of the outputs
+    )
+    pset.execute(
+        runtime=np.timedelta64(10, "m"), dt=np.timedelta64(60, "s"), pyfunc=AdvectionEE, output_file=output_file
+    )
